backend/requirements.txt
Flask
flask_sqlalchemy.SQLAlchemy